coaches CRUD images

+++ controllers/coachesController.js
@@ -38,7 +38,6 @@ export const coachesController = {
                 club,
                 role,
                 phone,
-                ensurance,
                 pay,
                 active,
                 createdAt
@@ -47,12 +46,20 @@ export const coachesController = {
             let image;
 
             if (req.files?.image) {
+
                 const imagePosted = await uploadImage(req.files.image.tempFilePath)
+                
                 image = {
                     url: imagePosted.secure_url,
                     public_id: imagePosted.public_id
                 }
+
                 await fs.remove(req.files.image.tempFilePath)
+            } else {
+                image = {
+                    url: 'https://res.cloudinary.com/dlah9v2do/image/upload/v1680549252/FotosPerfil/opkjqvstjmumhgz2azvw.png',
+                    public_id: 'FotosPerfil/opkjqvstjmumhgz2azvw'
+                }
             }
 
             const newCoach = new Coach({
@@ -63,7 +70,6 @@ export const coachesController = {
                 club,
                 role,
                 phone,
-                ensurance,
                 pay,
                 active,
                 createdAt
@@ -87,7 +93,7 @@ export const coachesController = {
                 await deleteImage(coachToUpdate.image.public_id)
 
                 let newImage;
-                
+
                 const imageUpdate = await uploadImage(req.files.image.tempFilePath)
 
                 newImage = {
@@ -121,7 +127,7 @@ export const coachesController = {
         try {
             const coachRemoved = await Coach.findByIdAndDelete(req.params.id);
 
-            if (coachRemoved.image.public_id) {
+            if (coachRemoved.image.public_id && coachRemoved.image.url !== 'https://res.cloudinary.com/dlah9v2do/image/upload/v1680549252/FotosPerfil/opkjqvstjmumhgz2azvw.png') {
                 await deleteImage(coachRemoved.image.public_id)
             }
